Redefine Bet and create subclasses
Bet subclasses no longer store the type and bet amount (wager) as
instance variables. Instead, the type is set as a class variable.
The wager should be stored in a dictionary instead.

Each subclass of Bet defines the check() method, which is used to
determine the outcome of a bet after each roll.
Also add a BetOutcome enum to represent each possible outcome.