Remove Bet.can_add(), use Bet.max_wager() instead
If there is no wager for a bet, Bet.max_wager() can return 0 to signify
that a bet cannot be added. In other words, Createable implies
Increasable, and vice versa.

In contrast, Bet.can_remove() cannot be removed (zing!), because a bet
may have a minimum wager AND still be removable, or have no minimum
wager AND be unremovable (in which case, the effective minimum is 1).