Remove Bet.name, let the client choose names
The bet view classes should not be concerned about providing a human-
readable name for each bet type. Also, BetType.to_class() was supposed
to be an internal utility method for the server; it shouldn't be used by
the client to retrieve a human-readable string.

Let the client figure out what strings to show to the player.